requirements-test.txt
pytest>=8.0
pytest-mock>=3.12
apache-airflow>=3.2
